Extract message handler telemetry module
